add sliderfloat3 for model transformation. Todo : link it with cbuffer

+++ GraphicsPortfolio1/GraphicsPortfolio1/IMeshGroup.cpp
@@ -47,7 +47,7 @@ void IMeshGroup::Render(ComPtr<ID3D11DeviceContext>& device_context)
 		vector<ID3D11Buffer*> vertex_constant_buffers{ mesh->m_vertex_cbuffer_.Get(), m_vertex_stage_cbuffer_.Get() };
 		//vector<ID3D11Buffer**> pixel_constant_buffers{ mesh->m_vertex_cbuffer_.GetAddressOf(), m_vertex_cbuffer_.GetAddressOf() };
 
-		device_context->VSSetConstantBuffers(0, vertex_constant_buffers.size(), vertex_constant_buffers.data());
+		device_context->VSSetConstantBuffers(0, (UINT)vertex_constant_buffers.size(), vertex_constant_buffers.data());
 		//device_context->PSSetConstantBuffers(0, 1, mesh->m_pixel_cbuffer_.GetAddressOf());
 
 		device_context->PSSetShaderResources(0, 1, mesh->m_texture_sr_view_.GetAddressOf());
